Исправлен сундук, добавлено немного нового текста. Добавлена модель люка в затопленную часть. Исправлена случайная анимация бега у босса.

+++ Assets/Scripts/InventorySlotUI.cs
@@ -78,8 +78,10 @@ public class InventorySlotUI : MonoBehaviour, IPointerClickHandler
         else if (eventData.button == PointerEventData.InputButton.Left && chest.gameObject.activeSelf)
         {
             InventoryItem itemCopy = new InventoryItem(item);
-            context.inventory.DeleteItem(item.inventorySlotID, 99);
-            chest.AddItem(itemCopy);
+            if (chest.AddItem(itemCopy))
+            {
+                context.inventory.DeleteItem(item.inventorySlotID, 99);
+            }
         }
     }
 }
